+++ src/schema/validators/sub-blocks.ts
@@ -6,6 +6,7 @@ export * from '../../sub-blocks/NewsCard/schema';
 export * from '../../sub-blocks/Partner/schema';
 export * from '../../sub-blocks/TutorialCard/schema';
 export * from '../../sub-blocks/CardWithImage/schema';
+export * from '../../sub-blocks/LayoutItem/schema';
 export * from '../../sub-blocks/Quote/schema';
 export * from '../../sub-blocks/Divider/schema';
 export * from '../../sub-blocks/BasicCard/schema';