+++ src/sub-blocks/HubspotForm/index.tsx
@@ -30,6 +30,7 @@ const HubspotForm: React.FunctionComponent<HubspotFormProps> = (props) => {
         onBeforeLoad,
         onLoad,
         inVirtualDom = true,
+        onSubmitError,
     } = props;
 
     const handleMetrika = useMetrika();
@@ -44,12 +45,13 @@ const HubspotForm: React.FunctionComponent<HubspotFormProps> = (props) => {
             onBeforeLoad,
             onBeforeSubmit,
             onLoad,
+            onSubmitError,
             onSubmit: (e) => {
                 handleMetrika?.({pixelEvents});
                 onSubmit?.(e);
             },
         }),
-        [onBeforeLoad, onBeforeSubmit, onLoad, handleMetrika, pixelEvents, onSubmit],
+        [onBeforeLoad, onBeforeSubmit, onLoad, handleMetrika, pixelEvents, onSubmit, onSubmitError],
     );
 
     useHandleHubspotEvents(handlers, formId);