+++ src/sub-blocks/LayoutItem/utils.ts
@@ -1,0 +1,14 @@
+import {LayoutItemProps, LinkTheme, MediaProps} from '../../models';
+
+export const getLayoutItemLinks = (links: LayoutItemProps['content']['links']) =>
+    links?.map((link) => ({
+        theme: 'normal' as LinkTheme,
+        ...link,
+    }));
+
+export const hasFullScreen = ({dataLens, image}: MediaProps) => {
+    // datalens and slider media card don't support fullScreen mode
+    return !(dataLens || Array.isArray(image));
+};
+
+export const showFullScreenIcon = ({youtube}: MediaProps) => !youtube;